refactor: arquitectura hexagonal
